NOJIRA: Refactor pcp
Change file structure; Split index.js into multiple files - settingsPanel, exemplars
