+++ src/pcp/html/message.html
@@ -21,7 +21,7 @@
         <div class="petal petal12"></div>
       </div>
     </div>
-    <span class="gpii-icon">GPII <img src="../icons/gear-cloud-black.png"></img></span>
+    <span class="gpii-icon">GPII <img src="../../icons/gear-cloud-black.png"></img></span>
   </div>
 </body>
 </html>
