+++ src/pcp/index.html
@@ -2,13 +2,15 @@
     <head>
         <link href="../../node_modules/infusion/dist/assets/src/framework/preferences/css/PrefsEditor.css" rel="stylesheet" type="text/css">
         <link rel="stylesheet" type="text/css" href="../../node_modules/infusion/dist/assets/src/components/switch/css/Switch.css">
-        <link rel="stylesheet" href="../css/widgets.css">
-        <link rel="stylesheet" href="../css/settings.css">
-        <link rel="stylesheet" href="../css/splash.css">
+        <link rel="stylesheet" href="css/widgets.css">
+        <link rel="stylesheet" href="css/main.css">
+        <link rel="stylesheet" href="css/splash.css">
         <script type="text/javascript" src="../../node_modules/infusion/src/lib/jquery/core/js/jquery.js"></script>
         <script type="text/javascript" src="../../node_modules/infusion/dist/infusion-all.js"></script>
-        <script type="text/javascript" src="../widgets.js"></script>
-        <script type="text/javascript" src="../settings.js"></script>
+        <script type="text/javascript" src="js/widgets.js"></script>
+        <script type="text/javascript" src="js/exemplars.js"></script>
+        <script type="text/javascript" src="js/settingsPanel.js"></script>
+        <script type="text/javascript" src="js/index.js"></script>
     </head>
     <body id="flc-body">
         <div class="flc-splash">
